API/run.sh: 'Fix' the random config replacement
Issue:
Sometimes testing took the local conf, sometimes web took the testing conf.
As I found out, that every time the `php artisan config:cache` is called then it replaces the previous version.
https://stackoverflow.com/a/29729639
In my opion that's crazy but it is what it is.

Solution:
So the workaround is not to
cache configs an let them be read not-from-cache fiels.